requirements.txt
langchain
langchain-core
langchain-openai
openai
langchain-anthropic
langchain-google-genai
google-generativeai
streamlit
langchain-huggingface
python-dotenv
huggingface-hub